requirements.txt
python-telegram-bot==20.4
pdfplumber==0.10.1
fastapi==0.103.0
uvicorn==0.23.1
